requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
Django==5.2
django-bootstrap4==25.1
django-storages==1.14.6
libusb1==3.3.1
ndeflib==0.3.3
nfcpy==1.0.4
pillow==11.1.0
pyDes==2.0.1
pyserial==3.5
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.13.1
